backend/requirements.txt
Django==6.0.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1

# Production
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1